requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Templates & Forms
jinja2==3.1.2
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg[binary,pool]==3.1.18

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography>=41.0.0
crypto
pycrypto
pycryptodome

# Vector Database & Search
qdrant-client==1.17.0

# AI / Embeddings
google-genai
openai
anthropic
tiktoken

# LangChain / LangGraph multi-agent
langchain==0.3.7
langchain-core==0.3.21
langchain-google-genai==2.0.6
langchain-openai==0.2.10
langchain-anthropic==0.3.0
langgraph==0.2.53
langgraph-checkpoint-postgres==2.0.8
numpy==1.26.4
beautifulsoup4==4.12.3

# Server-side export (chat history PDF)
weasyprint==62.3

# Data Validation
pydantic==2.5.0

# Caching
redis==5.0.1

# HTTP Client
requests==2.31.0
aiohttp==3.9.1

# Environment Management
python-dotenv==1.0.0

# Development & Testing (optional but useful)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pydantic[email]